+++ src/initialize.py
@@ -2,8 +2,8 @@ from evdev import UInput, ecodes as e, AbsInfo
 import yaml
 import os
 
-async def initialize(scriptData: dict) -> tuple:
-    finalDevice = getFinalDevice(scriptData)
+def initialize() -> tuple:
+    finalDevice = getFinalDevice('taskZen-minimal')
     if finalDevice is None:
         return None, 'Device not found'
 
@@ -16,54 +16,31 @@ async def initialize(scriptData: dict) -> tuple:
         number = allKeys.get(key)
         keyList.append(number)
 
-    screenWidth = scriptData.get('screen', {}).get('width', 1920)
-    screenHeight = scriptData.get('screen', {}).get('height', 1080)
-    # Create the capabilities dictionary
+    # Create capabilities
     cap = {
         e.EV_KEY: keyList,
-        e.EV_ABS: [
-            (e.ABS_X, AbsInfo(value=0, min=0, max=screenWidth, fuzz=0, flat=0, resolution=0)),
-            (e.ABS_Y, AbsInfo(value=0, min=0, max=screenHeight, fuzz=0, flat=0, resolution=0)),
-        ],
         e.EV_REL: [
             (e.REL_X),
-            (e.REL_Y),
+            (e.REL_Y)
         ],
+        e.EV_ABS: [
+            (e.ABS_X, AbsInfo(0, 0, 1920, 0, 0, 0)),
+            (e.ABS_Y, AbsInfo(0, 0, 1080, 0, 0, 0))
+        ]
     }
 
-    # Create the virtual input device with absolute positioning
-    ui = UInput(cap, name=finalDevice['name'], phys=finalDevice['phys'], version=finalDevice['version'])
+    device = UInput(cap, name=finalDevice['name'], phys=finalDevice['phys'])
+    return device
 
-    return ui, None
-
-def getFinalDevice(scriptData: dict) -> dict:
-    # Find all matching devices
+def getFinalDevice(deviceName: int) -> dict:
+    # Find the device
     deviceDir = os.getenv('XDG_CONFIG_HOME', default=os.path.expanduser('~/.config')) + '/taskZen/devices/'
-    foundDevices = []
     for file in os.listdir(deviceDir):
         with open(deviceDir + file, 'r') as f:
             deviceData = yaml.safe_load(f)
-        if scriptData['device'] == deviceData['name']:
-            foundDevices.append(file)
-
-    finalDevice = None
-    # Match the version
-    if scriptData.get('device-version') is None: # Get newest version
-        foundDevicesData = []
-        for device in foundDevices:
-            with open(deviceDir + device, 'r') as f:
-                device = yaml.safe_load(f)
-            foundDevicesData.append(device)
-        finalDevice = max(foundDevicesData, key=lambda x: x['version'])
-
-    else: # Get matching version
-        for device in foundDevices:
-            with open(deviceDir + device, 'r') as f:
-                deviceData = yaml.safe_load(f)
-            if deviceData['version'] == scriptData.get('device-version'):
+            if deviceName == deviceData['name']:
                 finalDevice = deviceData
-                break
-        
+
     return finalDevice
 
 def getAllKeys() -> dict:
@@ -109,33 +86,3 @@ def findScript(scriptName: str) -> str:
             data = yaml.safe_load(file)
             if data['name'] == scriptName:
                 return scriptDir + script
-
-def scriptContainsExec(scriptData: dict) -> bool:
-    """
-    Checks if the script contains an exec command, including within loops
-    
-    Parameters:
-        - scriptData (dict): The script data
-    
-    Returns:
-        - bool: Whether the script contains an exec command
-    """
-    def containsExec(steps: list) -> bool:
-        for step in steps:
-            if step['type'] == 'exec':
-                return True
-            elif step['type'] == 'loop' and 'subSteps' in step:
-                if containsExec(step['subSteps']):
-                    return True
-            elif step['type'] == 'if' and 'trueSteps' in step and 'falseSteps' in step:
-                if containsExec(step['trueSteps']) or containsExec(step['falseSteps']):
-                    return True
-            elif step['type'] == 'if' and 'trueSteps' in step:
-                if containsExec(step['trueSteps']):
-                    return True
-            elif step['type'] == 'if' and 'falseSteps' in step:
-                if containsExec(step['falseSteps']):
-                    return True                
-        return False
-
-    return containsExec(scriptData.get('steps', []))
