+++ src/taskZen/__init__.py
@@ -1,0 +1,8 @@
+from .actions import *
+import initialize
+
+ui = initialize.initialize()
+allKeys = initialize.getAllKeys()
+actions.setVariables(ui, allKeys)
+
+__all__ = ['actions']
